drift_studio/ddoc/plugins/ddoc-plugin-vision/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ddoc-plugin-vision"
version = "1.0.0"
description = "Vision analysis plugin for ddoc (image analysis, drift detection)"
requires-python = ">=3.9"
dependencies = [
    "pluggy>=1.5.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "clip @ git+https://github.com/openai/CLIP.git",
    "pillow>=9.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "scikit-learn>=1.0.0",
    "scikit-image>=0.19.0",
    "scipy>=1.7.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "pyyaml>=6.0",
    "yellowbrick>=1.5",
]

[project.entry-points."ddoc"]
ddoc_vision = "ddoc_plugin_vision.vision_impl:DDOCVisionPlugin"

[tool.setuptools]
packages = {find = {where = ["."], include = ["ddoc_plugin_vision*"]}}
include-package-data = true
zip-safe = false
